Added read string ability
This commit adds a new function called keyboard_read_str, which reads
in a string from the keyboard until a return character is encountered.
It echos all input to the console as well, so the user can see what
he or she is typing. In the future, this can be made optional.

Tested in main by echoing and during comparisons with strings.