Added trig functions
This adds implementations for sine, cosine, and tangent, as well
as M_PI. These will be needed for drawing curves in graphics. All
values are in radians.

The implementation of these functions uses a polynomial approximation
which is described at http://lab.polygonal.de/?p=205. Using an
approximation is both easier to implement and more performant than
calculating the actual taylor series. I tested this by comparing
the sin and cos implementation to the ones provided in libc over a
range of 0-2PI. They differed by at most 0.06, and averaged ~0.03.
I think this is a reasonable range given the benefits of this approach.