Added arc/circle drawing
This adds the ability to draw colored arcs and circles of varying radius
to the screen. This was tested by drawing circles of various sizes, one
of which I left in main as a proof of function. The functions use sin
and cos to determine the location of pixels to color as theta is
gradually ranged from start_t to end_t

Tested by drawing several circles on the screen.