Ensure pagetable_init direct maps only lower vaddr
Currently, pagetable_init direct maps all available memory, which is
incorrect because when we have >= 2GB of physical memory present, the
higher addresses should be indirectly mapped.

Tested by running the system under different amounts of ram and make
sure the system still boots properly. Because of the hang in VirtualBox
with >= 512MB of memory, we can't test whether *not mapping* the higher
addresses breaks anything just yet.