Added graphics testing framework
This adds a framework that will be used to do performance testing of the
graphics library, and potentially other frameworks in the future. Users
create functions that test their features, and the framework will
monitor the runtime of these functions to make sure that it is within
acceptable limits. To add a test to the test suite, simply add it to the
graphics_tests array in graphics_test.h