Implement EOF
Implements EOF markers for reading from a file. Checks to see if you
read past the end of a file, and if so, returns the EOF marker. 
Tested by creating an ISO Filesystem with genisoimage, mounting that as
an optical disk in VirtualBox as the secondary slave, opening the file
with iso_fopen("/PATH/TO/FILE", 3) [ata device 3 for secondary slave],
and reading one character at a time and printing to screen.