Added testing framework
This adds the ability to create functions to test our various modules.
Tests can be added by adding a new entry to the tests array in testing.h.
The framework will both confirm that the tests return true, indicating
that all of their checks passed, and that the test elapsed within
a given amount of time.