Removes Redundant is_dir Check
Removes a redundant check in iso_dopen that ensured the entity fetched
was actually a directory. It is redundant because that is now handled
within iso_recursive_look_up.
Tested by removing the redundant check and still able to successfully
open a directory/unsuccessfully attempt to open a file as a directory.