NUN-13 Marked pages as used in kmalloc
This changes kmalloc to get pages directly from the memory_raw module,
as opposed to the process pagetable. This allow us to mark the page
as used in the memory module, so that others do not try to use it and
it is not allocated to otehr areas.

Tested by creating a window, which uses kmalloc, and it still works.