Add Border and String drawing
This adds functions to draw borders of windows, and to draw strings in
windows. When a window is created, a default border color of grey
(128,128,128) is used. To change this, the window_set_border_color()
function must be used, so that the border can be redrawn in the new color.

Tested by creating a window and drawing text to it.