Windows with Parents
This adds nested windows. When a window performs a draw, the actual location
of that draw will be determined relative to the entire line of windows
it is derived from. Windows will also be clipped by their parents bounds.

This also adds a test that can be run at the command line, window_test,
which tests the relative drawing and clipping of nested windows.