Add Window System Calls
This adds a preliminary set of window system calls. All drawing is done
in the current process' window. If there is no window for the calling
process, then no draw calls will occur and an error will be returned.

Tested by creating a process based on the runproc example and drawing