Added Window events
This adds a preliminary implementation of the event system for windows.
Events will be captured and added to an event queue for the active
window. The process can retrieve these events in the order that they
came in using the get_event system call. They provide a buffer into
which the event information should be copied.

If there is no active window, events will be distributed directly to
the kernel console.