Added window clear
This adds a window clearing to system calls, enabling people to create cleaner
looking programs. This is accomplish using the graphics_rect function to set
the window area to the window's background color.

Tested by adding functionality to clear after a certain number of characters
have been entered in the window_test function