Fix guru bug, addcrossplatform build support
The guru bug was caused by a large static allocation in the mouse draw buffer. This caused kernel space to extend into the interrupt table and corrupt it, causing triple faults. To fix this (for now) we alloce space for this during mouse_init. This buffer size (30*30*8) is too large to be allocated with kmalloc, which does not allow for allocations across pages currently. So we use the direct memory_alloc_page instead. We call this method twice in order to reserve two adjacent pages. The adjacency of the allocations is not guranteed, but can be implied from the early level in which this code happens. Once we can allocate memory larger than a page safely, we should revisit this issue

    Also add the ability to build nunya on multiple filesystems by abstracting away commands used for building. You can override $CC, $LD, and $ISOGEN to define your own paths to the tools needed to build. Default values have been provided to ensure backwards compatability with older setups.