Fix command line memory corruption
Our previous handling of string tokenizing in the command line resulted in potentially corrupted memory being read as the second parameter to commands. This fixes that by checking to ensure that there is a second part of the string after tokenization, and otherwise specifically uses a null string

You can check this by running ls in succession before this change, where sometimes it will succeed and sometimes it will attempt to read a garbage file name.